Fix this

+++ src/main/kotlin/it/tn/spoilers/database/extensions/BooksExtension.kt
@@ -31,6 +31,6 @@ fun BooksData.toBooks(): Books =
         Book_School_Type = this.Book_School_Type,
         Book_SchoolCode = this.Book_SchoolCode,
         Book_Study_Year = this.Book_Study_Year,
-        Book_Volume = this.Book_Volume,
+        Book_Volume = this.Book_Volume.toInt(),
         Book_YearSelection = this.Book_YearSelection
     )
